refactor(convex-native): extract client-safe overbooking-core for native reads
The overbooked-status math (computeOverbookedStatus) was server-only because its
module imports getConvexClient. The native read layer needs to run the SAME math
client-side over a useQuery(api.overbooking.bundle) subscription.

Extract the pure pieces into src/lib/overbooking-core.ts (zero server imports):
computeStockBreakdown, OverbookedInfo, DateWindow, EXCLUDED_PROJECT_STATUSES,
projectMatchesWindow, indexProjectsById, sumBookingsByModel, and a new
reconstructOverbookedStatus (the body of computeOverbookedStatus, taking the
bundle payload instead of fetching it).

- availability.ts computeOverbookedStatus is now a thin IO wrapper: fetch the
  bundle, delegate to reconstructOverbookedStatus (parity-by-construction).
- availability.ts re-exports computeStockBreakdown + OverbookedInfo; availability-
  read.ts re-exports the moved window/booking helpers — every existing server
  caller keeps its import path.
- The only value dependency is the already-client-safe mapLineItemDoc.

Adds overbooking-core.test.ts (8 cases): overbook, within-stock, reducedOnly,
cross-project bookings, sub-hire exclusion, dateless, kit-parent inheritance.
